refactor service, add request, methods in controller
