--- src/main/java/com/glinka/wcn/model/converter/daoToDto/UserDaoToDtoConverter.java
@@ -1,26 +1,0 @@
-package com.glinka.wcn.model.converter.daoToDto;
-
-import com.glinka.wcn.model.converter.ConverterAdapter;
-import com.glinka.wcn.model.dao.UserDao;
-import com.glinka.wcn.model.dto.User;
-import org.springframework.stereotype.Component;
-
-@Component
-public class UserDaoToDtoConverter extends ConverterAdapter<User, UserDao> {
-
-    @Override
-    public User convert(User target, UserDao source) {
-
-        if (target == null || source == null)
-            return null;
-
-        target.setId(source.getId());
-        target.setName(source.getName());
-        target.setSurname(source.getSurname());
-        target.setEmail(source.getEmail());
-        //TODO encrypt password
-        target.setPassword(source.getPassword());
-
-        return target;
-    }
-}
