--- src/main/java/com/glinka/wcn/request/category.http
@@ -1,13 +1,0 @@
-POST http://localhost:8282/saveCategory
-Content-Type: application/json
-
-{
-  "id": 2,
-  "name": "category1"
-}
-
-###
-GET http://localhost:8282/allCategory
-Accept: application/json
-
-###
