--- src/main/java/com/glinka/wcn/request/journal.http
@@ -1,26 +1,0 @@
-POST http://localhost:8282/saveScientificJournal
-Content-Type: application/json
-
-{
-  "id": 3,
-  "title1": "tyt1",
-  "issn1": "i1",
-  "eissn1": "e1",
-  "title2": "tyt2",
-  "issn2":  "i2",
-  "eissn2": "e2",
-  "points": 50,
-  "categories": [
-    {
-      "id": 2,
-      "name": "category1"
-    }
-  ]
-}
-
-###
-GET http://localhost:8282/allScientificJournal
-Accept: application/json
-
-###
-
