--- src/main/java/com/glinka/wcn/request/user.http
@@ -1,21 +1,0 @@
-POST http://localhost:8282/saveUser
-Content-Type: application/json
-
-{
-  "id": 1,
-  "name": "Jakub",
-  "surname": "Glinka",
-  "email": "jakub@mail.com",
-  "password": "haslo"
-}
-###
-
-GET http://localhost:8282/allUsers
-Accept: application/json
-
-###
-
-GET http://localhost:8282/findUserById?id=1
-Accept: application/json
-
-###
