+++ src/main/java/com/glinka/wcn/service/CategoryService.java
@@ -1,6 +1,6 @@
 package com.glinka.wcn.service;
 
-import com.glinka.wcn.model.dao.CategoryDao;
+import com.glinka.wcn.commons.ResourceNotFoundException;
 import com.glinka.wcn.model.dto.Category;
 
 import java.util.List;
@@ -10,8 +10,8 @@ public interface CategoryService {
 
     List<Category> findAll();
     List<Category> findAllById(List<Integer> ids);
-    List<CategoryDao> findAllDaoById(List<Integer> ids);
-    Category findById(int id);
-    boolean save(Category category);
+    Category findById(Integer id) throws ResourceNotFoundException;
+    Category save(Category category);
+    void delete(Integer id) throws ResourceNotFoundException;
 
 }
