+++ src/main/java/com/glinka/wcn/service/UserService.java
@@ -1,7 +1,6 @@
 package com.glinka.wcn.service;
 
 import com.glinka.wcn.commons.ResourceNotFoundException;
-import com.glinka.wcn.model.dao.UserDao;
 import com.glinka.wcn.model.dto.User;
 
 import java.util.List;
@@ -10,21 +9,10 @@ import java.util.List;
 public interface UserService {
 
     List<User> findAll();
-
     List<User> findAllById(List<Integer> ids);
-    List<UserDao> findAllDaoById(List<Integer> ids);
-
-    //   Find user by id/name/surname/email
     List<User> findAllByNameOrSurname(String name);
-
-//    User findById(Integer id) throws ResourceNotFoundException;
-    User findById(int id);
+    User findById(Integer id) throws ResourceNotFoundException;
     User findByEmail(String email);
-
-    //   Add new User
-    boolean save(User user);
-//    User save(User user);
-
-//    void delete(Integer id);
-//    void delete(Integer id) throws ResourceNotFoundException;
+    User save(User user);
+    void delete(Integer id) throws ResourceNotFoundException;
 }
